Add function to change grid from user input
Finished adding an event listener to the button to allow the user to
resize the grid based on their input. Also cleaned up the original code
to allow the grid to fill the container properly each time it is remade.
Put all of the original grid generating code into a function to reuse it
every time a grid is made or remade.